test: merge emergence tests into test_count.py
Move emergence detection tests into test_count.py to match the
source code refactor, removing the standalone test_emergence.py.
